extract TrackView component from TrackList
