+++ app/(client)/tracks/components/TrackList.tsx
@@ -1,21 +1,12 @@
 'use client';
-
-import Image from 'next/image';
-import Link from 'next/link';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { fab } from '@fortawesome/free-brands-svg-icons';
 import { motion, AnimatePresence } from 'framer-motion';
+import TrackView from './TrackView';
 
 interface Props {
   tracks: SpotifyApi.TrackObjectFull[] | undefined;
 }
 
 export default function TrackList({ tracks }: Props) {
-  const getArtistString = (artists: SpotifyApi.ArtistObjectSimplified[]) => {
-    const artistNames = artists.map((artist) => artist.name);
-    return artistNames.join(', ');
-  };
-
   const variants = {
     visible: (i: number) => ({
       opacity: 1,
@@ -38,28 +29,7 @@ export default function TrackList({ tracks }: Props) {
             animate="visible"
             custom={index}
           >
-            <p className="font-bold text-md self-center">{index + 1}</p>
-            <Image
-              src={track.album.images[1].url}
-              alt="a track image"
-              width={50}
-              height={50}
-              className="w-20 h-20"
-            />
-            <div className="flex justify-between w-full gap-2 items-center">
-              <div className="flex flex-col">
-                <p className="text-md md:text-xl font-semibold">{track.name}</p>
-                <p className="text-sm md:text-lg font-extralight">
-                  {getArtistString(track.artists)}
-                </p>
-                <p className="text-sm md:text-lg font-extralight">
-                  BPM/Tempo: {Math.round(track.tempo)}
-                </p>
-              </div>
-              <Link href={track.uri} target="_blank">
-                <FontAwesomeIcon className={'h-8'} icon={fab.faSpotify} />
-              </Link>
-            </div>
+            <TrackView track={track} index={index} />
           </motion.li>
         ))}
       </ul>
